a widened view of a private member cannot write to it
the write requirement was computed against the class's own view, so the `T` an
enclosing class body holds satisfied it — `a = A[object](); a.t = t` passed even
though `a`'s parameter has nothing to do with `self`'s. a write is now the
erasure's bottom materialization, the dual of the erased read
